Creazione dto internalNotification Refactor del codice Sviluppo parziale 2.01 Creazione notifica

+++ src/main/java/it/pagopa/pn/delivery/svc/StatusService.java
@@ -2,7 +2,6 @@ package it.pagopa.pn.delivery.svc;
 
 import it.pagopa.pn.commons.exceptions.PnInternalException;
 import it.pagopa.pn.delivery.generated.openapi.msclient.datavault.v1.model.RecipientType;
-import it.pagopa.pn.delivery.generated.openapi.server.v1.dto.NotificationRecipient;
 import it.pagopa.pn.delivery.generated.openapi.server.v1.dto.NotificationStatus;
 import it.pagopa.pn.delivery.generated.openapi.server.v1.dto.RequestUpdateStatusDto;
 import it.pagopa.pn.delivery.middleware.NotificationDao;
@@ -13,11 +12,13 @@ import it.pagopa.pn.delivery.middleware.notificationdao.entities.NotificationCos
 import it.pagopa.pn.delivery.middleware.notificationdao.entities.NotificationDelegationMetadataEntity;
 import it.pagopa.pn.delivery.middleware.notificationdao.entities.NotificationMetadataEntity;
 import it.pagopa.pn.delivery.models.InternalNotification;
+import it.pagopa.pn.delivery.models.internal.notification.NotificationRecipient;
 import it.pagopa.pn.delivery.pnclient.datavault.PnDataVaultClientImpl;
 import it.pagopa.pn.delivery.utils.DataUtils;
 import lombok.extern.slf4j.Slf4j;
 import org.jetbrains.annotations.NotNull;
 import org.springframework.stereotype.Service;
+import org.springframework.util.StringUtils;
 import software.amazon.awssdk.enhanced.dynamodb.Key;
 
 import java.time.OffsetDateTime;
@@ -51,6 +52,7 @@ public class StatusService {
     }
 
     public void updateStatus(RequestUpdateStatusDto dto) {
+        //TODO: DO NOT REGRESSION TESTS
         Optional<InternalNotification> notificationOptional = notificationDao.getNotificationByIun(dto.getIun());
 
         if (notificationOptional.isPresent()) {
@@ -66,17 +68,18 @@ public class StatusService {
                 }
                 case REFUSED ->
                         notification.getRecipients().stream()
-                                .filter(r -> Objects.nonNull(r.getPayment()))
-                                .forEach(r -> {
-                                    notificationCostEntityDao.deleteItem(NotificationCostEntity.builder()
-                                            .creditorTaxIdNoticeCode(r.getPayment().getCreditorTaxId() +"##"+ r.getPayment().getNoticeCode())
-                                            .build());
-                                    if(r.getPayment().getNoticeCodeAlternative() != null && !r.getPayment().getNoticeCodeAlternative().isEmpty())
+                                .filter(r -> Objects.nonNull(r.getPayments()))
+                                .forEach(r -> r.getPayments().forEach(notificationPaymentInfo -> {
+                                    if(StringUtils.hasText(notificationPaymentInfo.getNoticeCode())){
                                         notificationCostEntityDao.deleteItem(NotificationCostEntity.builder()
-                                                .creditorTaxIdNoticeCode(r.getPayment().getCreditorTaxId() +"##"+ r.getPayment().getNoticeCodeAlternative())
+                                                .creditorTaxIdNoticeCode(notificationPaymentInfo.getCreditorTaxId() +"##"+ notificationPaymentInfo.getNoticeCode())
                                                 .build());
-
-                                });
+                                    }
+                                    if(StringUtils.hasText(notificationPaymentInfo.getNoticeCodeAlternative()))
+                                        notificationCostEntityDao.deleteItem(NotificationCostEntity.builder()
+                                                .creditorTaxIdNoticeCode(notificationPaymentInfo.getCreditorTaxId() +"##"+ notificationPaymentInfo.getNoticeCodeAlternative())
+                                                .build());
+                                }));
                 default -> {
                     Key key = Key.builder()
                             .partitionValue( notification.getIun() + "##" + notification.getRecipients().get( 0 ).getInternalId() )
@@ -117,8 +120,8 @@ public class StatusService {
         }
 
         return opaqueTaxIds.stream()
-                    .map( recipientId -> this.buildOneSearchMetadataEntry( notification, lastStatus, recipientId, opaqueTaxIds, creationMonth, acceptedAt))
-                    .toList();
+                .map( recipientId -> this.buildOneSearchMetadataEntry( notification, lastStatus, recipientId, opaqueTaxIds, creationMonth, acceptedAt))
+                .toList();
     }
 
     private NotificationMetadataEntity buildOneSearchMetadataEntry(
